Split `iroha_data_model::events` to files
